pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillful-mcp"
version = "0.1.0"
description = "MCP server with JWT-based skill authorization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "typer>=0.12",
    "semver>=3.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "anthropic>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
mcp-server = "mcp_server.main:run"
mcp-cli = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_server*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
